// models/ConferenceModel.ts
import { IContentItem, Elements } from '@kentico/kontent-delivery';

/**
 * Generated by '@kentico/kontent-model-generator@4.1.0'
 */
export type ConferenceModel = IContentItem<{
  urlSlugHistory: Elements.CustomElement;
  urlSlug: Elements.UrlSlugElement;
  title: Elements.TextElement;
  text: Elements.TextElement;
}>;