// models/_project.ts
/**
 * Generated by '@kentico/kontent-model-generator@4.1.0'
 */
export const projectModel = {
  languages: {
    default: {
      codename: 'default',
      name: 'Default project language',
    },
  },
  contentTypes: {
    conference: {
      codename: 'conference',
      name: 'Conference',
      elements: {
        url_slug_history: {
          codename: 'url_slug_history',
          name: 'URL slug history',
        },
        url_slug: {
          codename: 'url_slug',
          name: 'URL slug',
        },
        title: {
          codename: 'title',
          name: 'Title',
        },
        text: {
          codename: 'text',
          name: 'Text',
        },
      },
    },
    hero: {
      codename: 'hero',
      name: 'Hero',
      elements: {
        cta_url: {
          codename: 'cta_url',
          name: 'CTA URL',
        },
        headline: {
          codename: 'headline',
          name: 'Headline',
        },
        cta_label: {
          codename: 'cta_label',
          name: 'CTA label',
        },
        summary: {
          codename: 'summary',
          name: 'Summary',
        },
      },
    },
  },
  taxonomies: {},
};